Add make diagnosis feature
Creation of diagnosis events for a patient.

Use AJAX for diagnose form

Add validation to Diagnosis Events.

Make diagnosis needs to be sure that a diagnosis id is passed
so that it is able to copy the diagnosis into it's event.

Create a custom form builder to clean up the long grouped
collection select necessary for diagnosis selection.